+++ components/displayMessages/displayMessages.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Message, SignatureData } from '../../lib/types'
-import { StyledDivider, StyledErrorMessage } from '../../pageStyles/emailSender.styles'
-import { StyledFinalMessagesContainer } from './displayMessages.styles'
+import { StyledErrorMessage } from '../../pageStyles/emailSender.styles'
+import { StyledFinalMessagesContainer, StyledDivider } from './displayMessages.styles'
 import FinalMessage from '../../components/finalMessage/finalMessage'
 import EmailSignature from '../emailSignature/emailSignature'
 import { nanoid } from 'nanoid'
