+++ components/sendEmails/sendEmails.tsx
@@ -16,10 +16,9 @@ import React from 'react'
 import { ResultMessage } from '../../lib/types'
 import {
   StyledDateTimeDiv,
-  StyledResultMessage,
-  StyledSubHeader
-} from '../../pageStyles/emailSender.styles'
-import { SectionContainer, StyledButton } from '../../styles/common'
+  StyledResultMessage
+} from './sendEmail.styles'
+import { SectionContainer, StyledButton, StyledSubHeader } from '../../styles/common'
 import { icons } from '../../styles/icons'
 
 type SendEmailsProps = {
